Move embedded-search and popup-search in old_components
