+++ libs/search-widget/src/old-components/embedded-search/EmbeddedSearch.scss
