+++ libs/search-widget/src/old-components/popup-search/PopupSearch.scss
